requirements.txt
sounddevice
numpy
pyside6
zeroconf
soundfile
streamdeck==0.9.3
Pillow
